fix: 클라이언트에서 이미지 삭제 및 업로드가 작동하지 않던 버그 수정
- 'use server'에서 정의된 removeImageIfNeeded와 uploadToSignedUrl가 클라이언트에서 실행되지 않던 문제 해결
- 해당 함수들을 클라이언트 전용 파일로 분리하여 정상 작동하도록 수정
- 기존 이미지 삭제 로그 및 업로드 fetch가 동작하지 않던 현상 해결
