+++ src/app/actions/storage.ts
@@ -61,36 +61,3 @@ export async function getSignedUploadUrl(
 
   return { ...data, path }; // signedUrl, path
 }
-
-/* getSignedUploadUrl한테 받은 서명 URL로 fetch 업로드 */
-export async function uploadToSignedUrl(
-  signedUrl: string,
-  contentType: string,
-  file: Blob
-): Promise<boolean> {
-  const res = await fetch(signedUrl, {
-    method: 'PUT',
-    headers: { 'Content-Type': contentType },
-    body: file,
-  });
-
-  return res.ok;
-}
-
-/* 기존 이미지 삭제 */
-export async function removeImageIfNeeded(url: string, bucket: StorageBucket): Promise<void> {
-  const publicPrefix = `${process.env.NEXT_PUBLIC_SUPABASE_URL}/storage/v1/object/public/`;
-  // 절대 경로인 경우 → prefix 제거
-  let relativePath = url.startsWith(publicPrefix) ? url.replace(publicPrefix, '') : url;
-
-  // 상대 경로인 경우 → profiles/ 등 bucket 접두사 제거
-  if (relativePath.startsWith(bucket + '/')) {
-    relativePath = relativePath.replace(`${bucket}/`, '');
-  }
-
-  const supabase = await createClient();
-  const { error } = await supabase.storage.from(bucket).remove([relativePath]);
-
-  if (error) console.warn('기존 이미지 삭제 실패:', error.message);
-  else console.log('기존 이미지 삭제 성공:', relativePath);
-}
